Updated the Controller to properly walk through the MCC structure.
The Controller now walks through the MCC structure properly. To ensure
it has time to complete, the controller service also uses basic scaling
with max instances set to 1.

The ads-task-handler was also updated to an F4 instance to ensure it
doesn't run out of memory processing the landing page report. This might
only be an issue on the first run due to the 30 days of data.

Change-Id: I444221c3ac1a7fe6a07801a116c7f94630d3eb0e