Changed the Lighthouse task handler to use PSI.
It turns out you can't run multiple instances of Lighthouse in a
process. This means the LH-Sevice is not going to work. The LH team
suggests using PSI instead.

The lh task handler now uses PSI instead of the local service. To do
this, we also need an API key, which was added to the config service.

The controller now URL-encodes the URL for auditing before passing it to
the LH task handler.

Change-Id: I6c6f6107fce870f298dd0580c941c25e20a7246f